Update README.md
This repository contains a step-by-step web scraping project completed as part of the IBM Data Science Professional Certificate. The goal of this mini-project was to extract stock price data from a live website using Python and BeautifulSoup.

Although small in scale, the project was hands-on, practical, and helped reinforce core skills related to web scraping, data collection, and DataFrame manipulation using pandas.